requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.19.1
anyio==4.2.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-lru==2.0.4
attrs==23.2.0
av==12.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cloup==3.0.5
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.1.1
cryptography==42.0.8
cycler==0.11.0
Cython==3.0.10
dataclasses-json==0.6.4
debugpy==1.8.1
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.109.2
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.12.4
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
glcontext==2.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.27.0
google-crc32c==1.5.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==4.20.1
gradio_client==0.11.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.62.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.3
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
isosurfaces==0.1.2
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
jq==1.7.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.1
langchain-anthropic==0.1.4
langchain-community==0.2.1
langchain-core==0.2.1
langchain-experimental==0.0.53
langchain-google-genai==0.0.9
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langsmith==0.1.63
lark==1.1.9
lazy_loader==0.3
livekit==0.11.1
livekit-agents==0.7.2
livekit-api==0.5.1
livekit-plugins-deepgram==0.5.1
livekit-plugins-elevenlabs==0.6.0
livekit-plugins-openai==0.6.0
livekit-plugins-silero==0.5.1
livekit-protocol==0.5.1
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
moderngl==5.10.0
moderngl-window==2.4.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.30.3
opencv-python==4.8.0.76
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.15
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.2.1
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pgvector==0.3.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
posthog==3.4.1
praw==7.7.1
prawcore==2.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
psycopg2==2.9.9
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pulsar-client==3.4.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.26.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
pyee==11.1.0
pyglet==2.0.16
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
pylibsrtp==0.10.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.3.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_43pxpbos3z/croot/pyzmq_1705605108344/work
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.1
schedule==1.2.1
scikit-image==0.21.0
scipy==1.11.2
screeninfo==0.8.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
simple-websocket==1.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
skia-pathops==0.8.0.post1
sniffio==1.3.0
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.27
srt==3.5.3
stack-data==0.6.3
starlette==0.36.3
svgelements==1.9.6
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
tifffile==2023.8.30
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
trimesh==3.23.5
typer==0.9.0
types-protobuf==4.25.0.20240417
types-python-dateutil==2.8.19.20240106
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
